refactor(msa): backend registry — one module per kind, thin generate.py
Thermo-nuclear review finding #1/#2: generate.py had grown to 444 lines holding
three MSA-generation backends dispatched by scattered `if kind ==` branches in
prepare/submit/ingest. Extract each backend to msa/backends/<kind>.py with a uniform
prepare/submit/ingest interface + a BACKENDS registry; generate.py is now a ~45-line
dispatcher (unknown kinds raise clearly). Delete the dead legacy external
colabfold-local backend (_index_a3ms_by_hash/_split_complex/COLABFOLD_LOCAL) — the only
per_chain consumer was colabfold, whose runner ignores the store (fetches via
--msa-mode), so set colabfold msa: none. Behavior-preserving for every real run.

- msa/backends/{_common,boltz_csv,complex}.py + __init__ (BACKENDS)
- generate.py: registry + dispatch only
- registry/models.yaml: colabfold msa per_chain -> none
- tests: drop dead _split_complex test, add test_msa_backends_registered
